requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.32
pydantic==2.11.7
python-multipart==0.0.20
PyJWT==2.9.0
